Refactoring domain to model
