+++ src/main/java/io/bearch/webapi/book_service/repository/BookRepository.java
@@ -1,6 +1,6 @@
 package io.bearch.webapi.book_service.repository;
 
-import io.bearch.webapi.book_service.domain.Book;
+import io.bearch.webapi.book_service.model.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
